perf: 简单优化代码 (#63 @Gitea)

+++ plusone-commons/src/main/java/xyz/zhouxy/plusone/commons/util/EnumTools.java
@@ -154,7 +154,7 @@ public final class EnumTools {
      */
     @Nullable
     public static <E extends Enum<?>> Integer checkOrdinalNullable(Class<E> enumType, @Nullable Integer ordinal) {
-        return checkOrdinalOrDefault(enumType, ordinal, (Integer) null);
+        return checkOrdinalOrDefault(enumType, ordinal, null);
     }
 
     /**
